Add tools from toolbox-root
If we're making a custom image anyway, let's include all the tools we
need instead of pulling a toolbox just for some commands that we need as
root.

Ideally we would use LinuxBrew or Nix for those tools to avoid layering
but I haven't found a good setup for that yet.